レビュー指摘修正
- DTOの構成見直し
