+++ iplass-googlecloud/src/main/java/org/iplass/mtp/pushnotification/fcmv1/PushNotificationTarget.java
@@ -19,23 +19,71 @@
  */
 package org.iplass.mtp.pushnotification.fcmv1;
 
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 /**
- * FCM v1 API 用 Push 通知対象インターフェース
+ * FCM v1 API 用 Push 通知対象
  *
  * @author SEKIGUCHI Naoya
  */
-public interface PushNotificationTarget {
+public class PushNotificationTarget {
+	/** 対象タイプ */
+	private PushNotificationTargetType type;
+	/** 対象識別子 */
+	private String target;
+
+	/**
+	 * コンストラクタ
+	 * @param type 対象タイプ
+	 * @param target 通知対象
+	 */
+	public PushNotificationTarget(PushNotificationTargetType type, String target) {
+		this.type = type;
+		this.target = target;
+	}
 
 	/**
 	 * 対象タイプを取得します
 	 * @return 対象タイプ
 	 */
-	PushNotificationTargetType getType();
+	public PushNotificationTargetType getType() {
+		return type;
+	}
 
 	/**
 	 * 通知対象を取得します
 	 * @return 通知対象
 	 */
-	String getTarget();
+	public String getTarget() {
+		return target;
+	}
 
+	@Override
+	public String toString() {
+		return PushNotificationTarget.class.getSimpleName() + "{type=" + type + ", target=" + target + "}";
+	}
+
+	/** トークン、トピック、条件のいずれかのプレフィックス付き宛先のパターン */
+	private static final Pattern ANY_PREFIXED_VALUE_PATTERN = Pattern
+			.compile("^("
+					+ PushNotificationTargetType.TOKEN.getFieldName() + "|"
+					+ PushNotificationTargetType.TOPIC.getFieldName() + "|"
+					+ PushNotificationTargetType.CONDITION.getFieldName() + "):(.*)");
+
+	/**
+	 * 通知対象インスタンスを作成する
+	 * @param anyPrefixedValue トークン、トピック、条件のいずれかのプレフィックス付き宛先
+	 * @return 通知対象
+	 */
+	public static PushNotificationTarget create(String anyPrefixedValue) {
+		Matcher matcher = ANY_PREFIXED_VALUE_PATTERN.matcher(anyPrefixedValue);
+		if (!matcher.find()) {
+			throw new IllegalArgumentException("The target specified does not match the pattern. (target = " + anyPrefixedValue + ")");
+		}
+
+		String prefixIsFieldName = matcher.group(1);
+		String target = matcher.group(2);
+		return new PushNotificationTarget(PushNotificationTargetType.getTargetType(prefixIsFieldName), target);
+	}
 }
